Add files via upload
ChocoCrunch Analytics is a data analytics and visualization project focused on chocolate products. 
The project extracts real-time data from the OpenFoodFacts API, performs data cleaning, transformation, 
and feature engineering using Python, and stores it in a MySQL database. 

It introduces derived features such as sugar-to-carbohydrate ratio, calorie category, sugar category, 
and ultra-processed classification (based on NOVA group). 

An interactive Streamlit dashboard is developed to explore key insights such as:
- Distribution of calorie and sugar categories,
- Nutrient correlations,
- Top chocolate brands by product count,
- Relationship between calories and sugars, and
- Ratio of ultra-processed vs non-processed chocolates.

The main goal of the project is to demonstrate end-to-end data analytics — 
from API-based data extraction to database storage and real-time visualization using Streamlit.